Add Recursive. Extract Reading logic into a module.
